Hotfix of jumping in build mode
WARNING:

DO NOT consider this hotfix as part of the game jam, because this was done after the jam ended. I am only fixing this because I want to release the game publicly, but to do so, I need to fix the jump for all exe. files to correctly work